# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event-driven Notification API",
        "description": "API for creating, tracking, and validating notifications.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "queued",
                                "sending",
                                "accepted",
                                "delivered",
                                "failed",
                                "canceled",
                                "scheduled",
                                "unknown"
                            ]
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sms",
                                "email",
                                "push"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "high",
                                "normal",
                                "low"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paged notifications list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a notification",
                "operationId": "5e12e1940735be8dea4ffa6c52938048",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Optional key to ensure idempotent requests.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreNotificationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "batch_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/notifications/{notification}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get notification by id",
                "operationId": "ea7eaa587ff4cfb9b97294857f2adbff",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found."
                    }
                }
            }
        },
        "/api/notifications/batch/{batchId}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications for a batch",
                "operationId": "67442fa497bc44fc2dccf7c908069061",
                "parameters": [
                    {
                        "name": "batchId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch notifications.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notification"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
...
        "/api/notifications/{notification}/cancel": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Cancel a queued notification",
                "operationId": "e963b9faee81d2192b1e37ecdfac7f5a",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification canceled.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found."
                    },
                    "422": {
                        "description": "Notification cannot be canceled."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreNotificationRequest": {
                "description": "Provide either content or template_id + variables.",
                "required": [
                    "to",
                    "channel"
                ],
                "properties": {
                    "to": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email",
                            "push"
                        ]
                    },
                    "content": {
                        "type": "string",
                        "nullable": true
                    },
                    "template_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "variables": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "high",
                            "normal",
                            "low"
                        ]
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}